Make plugin-lists and MIDI-manager-list thread-safe
During debugging (PC) and testing (RaspberryPi) the MIDI import dialog, qtractor
crashed several times with segfaults due to inconsistent list state (the MIDI-
manager case was mentioned in [1]). Since adding plugins to MIDI import the
probabilty of these crashes increases because of performing 'dangerous'
operations multiple times.

To avoid this, lock/unlock guards were added under the following consideration:
Keep the locking time as short as possible to reduce the probability of
blocking audio-engine. For the implemetation qtractorSession::lock got a
parameter 'bStabilize' set true by default. If set to false, no events are
performed to avoid plugin-list-view updates with inconsistent data.

Worth to mention: With this modification no more crashes happend.

[1] https://github.com/rncbc/qtractor/issues/91

+++ src/qtractorPlugin.cpp
@@ -1620,10 +1620,17 @@ void qtractorPluginList::insertPlugin (
 	// We'll get prepared before plugging it in...
 	pPlugin->setChannels(m_iChannels);
 
+	qtractorSession *pSession = qtractorSession::getInstance();
+	if (pSession == NULL)
+		return;
+
+	// Do not perform events while locking - list views are not yet updated.
+	pSession->lock(false);
 	if (pNextPlugin)
 		insertBefore(pPlugin, pNextPlugin);
 	else
 		append(pPlugin);
+	pSession->unlock();
 
 	// Now update each observer list-view...
 	QListIterator<qtractorPluginListView *> iter(m_views);
@@ -1654,13 +1661,15 @@ void qtractorPluginList::movePlugin (
 	if (pPluginList == NULL)
 		return;
 
-	// Remove and insert back again...
+	qtractorSession *pSession = qtractorSession::getInstance();
+	if (pSession == NULL)
+		return;
+
+	// Do not perform events while locking - list views are not yet updated.
+	pSession->lock(false);
+	// Remove it -> it is no more processed by worker thread.
 	pPluginList->unlink(pPlugin);
-	if (pNextPlugin) {
-		insertBefore(pPlugin, pNextPlugin);
-	} else {
-		append(pPlugin);
-	}
+	pSession->unlock();
 
 	// DANGER: Gasp, we might be not the same...
 	if (pPluginList != this) {
@@ -1713,6 +1722,16 @@ void qtractorPluginList::movePlugin (
 		pListView->setCurrentItem(pNextItem);
 	}
 
+	// Views are updated so perform events during lock.
+	pSession->lock();
+	// Insert back again -> processing continues.
+	if (pNextPlugin) {
+		insertBefore(pPlugin, pNextPlugin);
+	} else {
+		append(pPlugin);
+	}
+	pSession->unlock();
+
 	// update (both) lists for Auto-plugin-deactivation
 	autoDeactivatePlugins(m_bAutoDeactivated, true);
 	if (pPluginList != this)
@@ -1723,8 +1742,15 @@ void qtractorPluginList::movePlugin (
 // Remove-guarded plugin method.
 void qtractorPluginList::removePlugin ( qtractorPlugin *pPlugin )
 {
-	// Just unlink the plugin from the list...
+	qtractorSession *pSession = qtractorSession::getInstance();
+	if (pSession == NULL)
+		return;
+
+	// Views are not touched here so perform events.
+	pSession->lock();
+	// Remove it -> it is no more processed by worker thread.
 	unlink(pPlugin);
+	pSession->unlock();
 
 	if (pPlugin->isActivated())
 		updateActivated(false);
